checkout confirm

+++ src/app/api/cart/[id]/route.js
@@ -11,9 +11,18 @@ export async function GET(req, {params}) {
 
 export async function DELETE(req, {params}) {
     const {id} = await params
-
-    const result = await cart.deleteOne({_id: new ObjectId(id)})
-    return Response.json(result)
+    const {searchParams} = new URL(req.url)
+    const userId = searchParams.get('userId')
+    console.log(userId);
+    
+    if (userId) {
+        const result = await cart.deleteMany({userID: new ObjectId(userId)})
+        return Response.json(result)
+    }
+    if (id) {
+        const result = await cart.deleteOne({_id: new ObjectId(id)})
+        return Response.json(result)
+    }
 }
 
 export async function POST(req, {params}) {
@@ -34,4 +43,16 @@ export async function POST(req, {params}) {
     }
     const result = await cart.insertOne(cartProduct)
     return Response.json({result, message: 'Cart added successfully', status: 201})
+}
+
+export async function PATCH(req, { params }) {
+  const { id } = await params;
+  const { quantity } = await req.json();
+
+  const result = await cart.updateOne(
+    { _id: new ObjectId(id) },
+    { $set: { quantity } }
+  );
+
+  return Response.json(result);
 }
